Cleaned up includes in libraries. Removed unnecessary dependency on types in testing library.

+++ libraries/testing/src/testing.cpp
@@ -8,9 +8,6 @@
 //
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
-// types
-#include "types.h"
-
 // testing
 #include "testing.h"
 
@@ -34,13 +31,13 @@ namespace testing
     bool isEqual(const std::vector<double>& a, const std::vector<double>& b, double tolerance)
     {
         // allow vectors of different size, provided that they differ by a suffix of zeros
-        uint64 size = a.size();
+        size_t size = a.size();
         if (b.size() < size)
         {
             size = b.size();
         }
 
-        for (uint64 i = 0; i < size; ++i)
+        for (size_t i = 0; i < size; ++i)
         {
             if (isEqual(a[i], b[i], tolerance) == false)
             {
@@ -49,7 +46,7 @@ namespace testing
         }
 
         // confirm suffix of zeros
-        for (uint64 i = size; i < a.size(); ++i)
+        for (size_t i = size; i < a.size(); ++i)
         {
             if (isEqual(a[i], 0, tolerance) == false)
             {
@@ -57,7 +54,7 @@ namespace testing
             }
         }
 
-        for (uint64 i = size; i < b.size(); ++i)
+        for (size_t i = size; i < b.size(); ++i)
         {
             if (isEqual(b[i], 0, tolerance) == false)
             {
